Id from int to long

+++ ApplicationCore/EntityBase.cs
@@ -2,7 +2,6 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Web.Mvc;
 
 namespace VTT.Data
 {
@@ -10,7 +9,7 @@ namespace VTT.Data
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [HiddenInput]
         public DateTime CreationDate { get; set; }
